extract CLI unit test function from renderer
